+++ src/core/mod.rs
@@ -2,13 +2,10 @@ use chrono::Utc;
 
 pub mod connectivity;
 pub mod exchanges;
-pub mod local_order_book_snapshot;
-pub mod local_snapshot_service;
 pub mod logger;
 pub mod orders;
 
-pub mod order_book_data;
-pub mod order_book_event;
+pub mod order_book;
 pub mod settings;
 
 pub type DateTime = chrono::DateTime<Utc>;
