+++ src/core/order_book/local_order_book_snapshot.rs
@@ -1,4 +1,4 @@
-use crate::core::order_book_data::OrderBookData;
+use crate::core::order_book::order_book_data::OrderBookData;
 use crate::core::DateTime;
 use rust_decimal::prelude::*;
 use std::collections::BTreeMap;
