+++ src/core/order_book/order_book_event.rs
@@ -1,5 +1,5 @@
-use super::exchanges::common::*;
-use crate::core::order_book_data::OrderBookData;
+use crate::core::exchanges::common::*;
+use crate::core::order_book::order_book_data::OrderBookData;
 use crate::core::DateTime;
 use derive_getters::Dissolve;
